requirements.txt
Pillow>=10.0.0
pynput>=1.7.6

